feat(lib): rename "new_track" to just "track"
Now that the old one is removed.
