+++ lib/src/lib.rs
@@ -10,12 +10,12 @@ pub mod config;
 pub mod ids;
 pub mod invidious;
 pub mod library_db;
-pub mod new_track;
 pub mod player;
 pub mod playlist;
 pub mod podcast;
 pub mod songtag;
 pub mod taskpool;
+pub mod track;
 pub mod types;
 pub mod ueberzug;
 pub mod utils;
