+++ tui/src/ui/components/database.rs
@@ -4,7 +4,7 @@ use termusiclib::config::SharedTuiSettings;
 use termusiclib::ids::Id;
 use termusiclib::library_db::const_unknown::{UNKNOWN_ARTIST, UNKNOWN_FILE, UNKNOWN_TITLE};
 use termusiclib::library_db::{Indexable, SearchCriteria, TrackDB};
-use termusiclib::new_track::DurationFmtShort;
+use termusiclib::track::DurationFmtShort;
 use termusiclib::types::{DBMsg, GSMsg, Msg};
 use termusiclib::utils::{is_playlist, playlist_get_vec};
 use tui_realm_stdlib::List;
