+++ tui/src/ui/components/progress.rs
@@ -2,8 +2,8 @@ use crate::ui::Model;
 use std::time::Duration;
 use termusiclib::config::TuiOverlay;
 use termusiclib::ids::Id;
-use termusiclib::new_track::DurationFmtShort;
-use termusiclib::new_track::MediaTypesSimple;
+use termusiclib::track::DurationFmtShort;
+use termusiclib::track::MediaTypesSimple;
 use termusiclib::types::Msg;
 use tui_realm_stdlib::ProgressBar;
 use tuirealm::event::NoUserEvent;
